pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-socketio>=5.5.1",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gtts>=2.5.4",
    "gunicorn>=23.0.0",
    "langdetect>=1.0.9",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "werkzeug>=3.1.3",
    "gevent>=24.11.1",
    "gevent-websocket>=0.10.1",
    "eventlet>=0.39.1",
]
